# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企業分析エージェント API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/analysis": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Post Analysis",
        "operationId": "post_analysis_api_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{result_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis Result",
        "operationId": "get_analysis_result_api_analysis__result_id__get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{result_id}/download": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Download Analysis",
        "operationId": "download_analysis_api_analysis__result_id__download_get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Result Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "pdf",
                "docx"
              ],
              "type": "string",
              "default": "pdf",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}/runs": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get Company Runs",
        "operationId": "get_company_runs_api_companies__company_id__runs_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{result_id}/share": {
      "post": {
        "tags": [
          "share"
        ],
        "summary": "Create Share",
        "operationId": "create_share_api_analysis__result_id__share_post",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Share Api Analysis  Result Id  Share Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/share/{share_id}": {
      "get": {
        "tags": [
          "share"
        ],
        "summary": "Get Shared Result",
        "operationId": "get_shared_result_api_share__share_id__get",
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}/deep-research": {
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Post Deep Research",
        "operationId": "post_deep_research_api_companies__company_id__deep_research_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Post Deep Research Api Companies  Company Id  Deep Research Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search Company",
        "operationId": "search_company_api_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare": {
      "post": {
        "tags": [
          "compare"
        ],
        "summary": "Post Compare",
        "operationId": "post_compare_api_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisRequest": {
        "properties": {
          "company_url": {
            "type": "string",
            "minLength": 1,
            "pattern": "^https?://.*",
            "title": "Company Url",
            "description": "分析対象の企業URL（例: https://www.toyota.co.jp/）"
          },
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "description": "キャッシュを無視して再分析する",
            "default": false
          },
          "template": {
            "type": "string",
            "enum": [
              "general",
              "job_hunting",
              "investment",
              "competitor",
              "partnership"
            ],
            "title": "Template",
            "description": "分析テンプレート",
            "default": "general"
          }
        },
        "type": "object",
        "required": [
          "company_url"
        ],
        "title": "AnalysisRequest"
      },
      "AnalysisResponse": {
        "properties": {
          "company_url": {
            "type": "string",
            "title": "Company Url"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "is_cached": {
            "type": "boolean",
            "title": "Is Cached",
            "default": false
          },
          "analyzed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed At"
          },
          "structured": {
            "$ref": "#/components/schemas/StructuredData"
          },
          "summary": {
            "$ref": "#/components/schemas/SummaryData"
          },
          "scores": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreData"
              },
              {
                "type": "null"
              }
            ]
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceInfo"
            },
            "type": "array",
            "title": "Sources"
          },
          "raw_sources": {
            "items": {
              "$ref": "#/components/schemas/RawSource"
            },
            "type": "array",
            "title": "Raw Sources"
          },
          "markdown_page": {
            "type": "string",
            "title": "Markdown Page",
            "default": ""
          },
          "diff_report": {
            "type": "string",
            "title": "Diff Report",
            "default": ""
          },
          "template": {
            "type": "string",
            "title": "Template",
            "default": "general"
          }
        },
        "type": "object",
        "required": [
          "company_url",
          "structured",
          "summary",
          "sources"
        ],
        "title": "AnalysisResponse"
      },
...
      "CompanyProfile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "founded": {
            "type": "string",
            "title": "Founded",
            "default": ""
          },
          "ceo": {
            "type": "string",
            "title": "Ceo",
            "default": ""
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          },
          "employees": {
            "type": "string",
            "title": "Employees",
            "default": ""
          },
          "capital": {
            "type": "string",
            "title": "Capital",
            "default": ""
          }
        },
        "type": "object",
        "title": "CompanyProfile"
      },
...
      "CompareRequest": {
        "properties": {
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 2,
            "title": "Urls"
          },
          "template": {
            "type": "string",
            "enum": [
              "general",
              "job_hunting",
              "investment",
              "competitor",
              "partnership"
            ],
            "title": "Template",
            "default": "general"
          }
        },
        "type": "object",
        "required": [
          "urls"
        ],
        "title": "CompareRequest"
      },
      "CompareResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/AnalysisResponse"
            },
            "type": "array",
            "title": "Results"
          },
          "comparison_summary": {
            "type": "string",
            "title": "Comparison Summary",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "CompareResponse"
      },
      "Financials": {
        "properties": {
          "revenue": {
            "type": "string",
            "title": "Revenue",
            "default": ""
          },
          "operating_income": {
            "type": "string",
            "title": "Operating Income",
            "default": ""
          },
          "net_income": {
            "type": "string",
            "title": "Net Income",
            "default": ""
          },
          "growth_rate": {
            "type": "string",
            "title": "Growth Rate",
            "default": ""
          }
        },
        "type": "object",
        "title": "Financials"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "runs"
        ],
        "title": "HistoryResponse"
      },
      "NewsItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "date": {
            "type": "string",
            "title": "Date",
            "default": ""
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "NewsItem"
      },
      "RawSource": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "その他"
          }
        },
        "type": "object",
        "required": [
          "url",
          "title",
          "content"
        ],
        "title": "RawSource"
      },
...
      "RiskItem": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "RiskItem"
      },
...
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "run_type": {
            "type": "string",
            "title": "Run Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "run_type",
          "status",
          "template",
          "started_at",
          "completed_at",
          "duration_ms",
          "error_message",
          "result_id"
        ],
        "title": "RunSummary"
      },
      "ScoreData": {
        "properties": {
          "financial_health": {
            "$ref": "#/components/schemas/ScoreItem"
          },
          "growth_potential": {
            "$ref": "#/components/schemas/ScoreItem"
          },
          "competitive_edge": {
            "$ref": "#/components/schemas/ScoreItem"
          },
          "risk_level": {
            "$ref": "#/components/schemas/ScoreItem"
          },
          "info_transparency": {
            "$ref": "#/components/schemas/ScoreItem"
          }
        },
        "type": "object",
        "title": "ScoreData"
      },
      "ScoreItem": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "default": 0
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "title": "ScoreItem"
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "SearchResponse"
      },
      "SearchResult": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "SearchResult"
      },
      "SourceInfo": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "その他"
          }
        },
        "type": "object",
        "required": [
          "url",
          "title"
        ],
        "title": "SourceInfo"
      },
...
      "StructuredData": {
        "properties": {
          "company_profile": {
            "$ref": "#/components/schemas/CompanyProfile"
          },
          "business_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Business Domains"
          },
          "products": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Products"
          },
          "financials": {
            "$ref": "#/components/schemas/Financials"
          },
          "news": {
            "items": {
              "$ref": "#/components/schemas/NewsItem"
            },
            "type": "array",
            "title": "News"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskItem"
            },
            "type": "array",
            "title": "Risks"
          }
        },
        "type": "object",
        "title": "StructuredData"
      },
      "SummaryData": {
        "properties": {
          "overview": {
            "type": "string",
            "title": "Overview",
            "default": ""
          },
          "business_model": {
            "type": "string",
            "title": "Business Model",
            "default": ""
          },
          "swot": {
            "$ref": "#/components/schemas/SwotAnalysis"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risks"
          },
          "competitors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Competitors"
          },
          "outlook": {
            "type": "string",
            "title": "Outlook",
            "default": ""
          }
        },
        "type": "object",
        "title": "SummaryData"
      },
      "SwotAnalysis": {
        "properties": {
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weaknesses"
          },
          "opportunities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Opportunities"
          },
          "threats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Threats"
          }
        },
        "type": "object",
        "title": "SwotAnalysis"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}